requirements.txt
addict
future
lmdb
numpy
opencv-python
Pillow
pyyaml
requests
scikit-image
scipy
tb-nightly
torch>=1.7
torchvision
tqdm
yapf
timm
thop
ptflops
onnx==1.15.0